added obliterate functionality to deleteEntries() in AbstractAtomCollection [Reviewed By: abuckner]

+++ trunk/src/main/java/org/atomserver/core/AbstractAtomCollection.java
@@ -131,6 +131,14 @@ abstract public class AbstractAtomCollection implements AtomCollection {
      */
     abstract protected EntryMetaData deleteEntry(EntryTarget entryTarget,
                                                  boolean setDeletedFlag) throws AtomServerException;
+    
+    /**
+     * The deleteEntry() method on the AtomCollection API delegates to this method within the subclass
+     * with the obliterate flag thrown 
+     * 
+     * @param entryDescriptor   basically the EntryTarget
+     */
+    abstract protected void obliterateEntry(EntryDescriptor entryDescriptor);
 
     // ----------
     //   statics
@@ -957,26 +965,33 @@ abstract public class AbstractAtomCollection implements AtomCollection {
                 new TransactionalTask<EntryMetaData>() {
                     public EntryMetaData execute() {
                         
-                        EntryMetaData entryMetaData =
+                        boolean isObliterate = (Boolean)QueryParam.obliterate.parse(request);
+                        
+                        //completely delete the entryTarget and return the metaData of that deleted target
+                        if(isObliterate){
+                            EntryMetaData lastMetaData = getEntry(entryTarget);
+                            obliterateEntry(entryTarget);
+                            return lastMetaData;
+                        }
+                        else{
+                            EntryMetaData entryMetaData =
                                 deleteEntry(entryTarget, setDeletedFlag());
 
-                        // Replace the XML file with a "deleted file"
-                        //  we wait to do this now that we know that the delete was successfull
-                        EntryMetaData entryMetaDataClone = (EntryMetaData) (entryMetaData.clone());
-                        int currentRevision = entryMetaData.getRevision();
-                        entryMetaDataClone.setRevision((currentRevision - 1));
+                            // Replace the XML file with a "deleted file"
+                            //  we wait to do this now that we know that the delete was successful
+                            EntryMetaData entryMetaDataClone = (EntryMetaData) (entryMetaData.clone());
+                            int currentRevision = entryMetaData.getRevision();
+                            entryMetaDataClone.setRevision((currentRevision - 1));
 
-                        getContentStorage().deleteContent(createDeletedEntryXML(entryMetaDataClone),
+                            getContentStorage().deleteContent(createDeletedEntryXML(entryMetaDataClone),
                                                           entryMetaData);
-                        return entryMetaData;
+                            return entryMetaData;
+                        }
                     }
                 }
         );
         return (entryMetaData == null) ? null : newEntry(abdera, entryMetaData, EntryType.link);
     }
-
-    //~~~~~~~~~~~~~~~~~~~~~~
-
     private boolean postProcessEntryContents(String entryXml, EntryMetaData entryMetaData) {
         log.debug("BEGIN AUTO_TAGGING................");
         EntryAutoTagger autoTagger = getAutoTagger();
